Unify the data type of graph to uint64_t

+++ globals.cuh
@@ -1,24 +1,26 @@
 #ifndef GLOBALS_CUH
 #define GLOBALS_CUH
+#include <cstdint>
+#include <climits>
 enum ALG_TYPE {
     BFS, SSSP, CC, PR
 };
-typedef uint SIZE_TYPE;
+typedef unsigned long long SIZE_TYPE;
 typedef unsigned long long EDGE_POINTER_TYPE;
 
 //typedef uint EDGE_POINTER_TYPE;
 
 struct EdgeWithWeight {
-    uint toNode;
-    uint weight;
+    unsigned long long toNode;
+    unsigned long long weight;
 };
 struct FragmentData {
-    uint startVertex = UINT_MAX - 1;
-    uint vertexNum = 0;
+    unsigned long long startVertex = UINT64_MAX - 1;
+    unsigned long long vertexNum = 0;
     bool isIn = false;
     bool isVisit = false;
     FragmentData() {
-        startVertex = UINT_MAX - 1;
+        startVertex = UINT64_MAX - 1;
         vertexNum = 0;
         isIn = false;
         isVisit = false;
